Create Program Exercise #8-2
Prompt:
Write a program called DistanceFromAverage which allows a user to enter up to 15 double values. The user should enter 99999 to quit entering numbers. Display an error message if the user quits without entering any numbers; otherwise, display a count of the numbers entered, the arithmetic average of the numbers, and each entered value and its distance from the average.